requirements.txt
asgiref==3.3.1
Django==3.0.4
psycopg2-binary>=2.8
djangorestframework==3.11.0
pytz==2020.4
sqlparse==0.4.1
requests==2.25.0
dj-database-url==0.5.0
python-decouple==3.3
drf-yasg==1.17.1
redis==3.4.1
django-redis==4.12.1